Fixed the git-cli didn't sync the rename file up problem
