+++ src/components/FooterEffect.jsx
